training/Classification/requirements.txt
apache-beam[gcp]==0.6.0
pillow==4.0.0
tensorflow==1.0.1
google-cloud-dataflow
google-cloud-storage

 